TODO: Need to validate inputs when calling sketch generators. Catches bug in main.rkt.
Sketch generators don't validate their `logical-inputs` argument. If you
pass in a list with too many elements, you will likely not get the
results you expect (which is what was happening to me). We need to
validate these inputs somehow, or make the input interface richer (i.e.
a hashmap with names)