Update DSP tests and comment most of them out
I've updated the DSP models to be more correct, and as a result, more of
these tests don't work. I'm not 100% sure if they should work or not --
the question is, are the Lattice and Intel DSPs able to function
combinationally? If they are, then these tests should work. If they're
not, then these tests should be updated to be sequential. In either case
we should move away from this style of testing and towards
end-to-end testing by calling main.rkt directly and examining the
results.