Only pass in inputs on first half of clock cycle (#360)
This is a pretty significant change. Conceptually, I realized this is
"more correct" while implementing #358.

Why is this more correct? Well, let's talk about the times when what we
currently have is *incorrect*: It's incorrect for one-cycle synthesis.
Currently, if we're synthesizing a one-cycle operation, then we eval
with inputs=inputs and clk=0, and we eval with inputs=inputs and clk=1,
and then we use that last eval to formulate our synthesis query. But
because inputs=inputs and not inputs=0 on the second eval, then the
combinational paths will have inputs and not 0. So this means that the
synthesis can choose to use the combinational paths, effectively
bypassing the need to use the registers in the one-cycle case!

(note that we don't actually use 0 but symbolic values/havoc in the
synthesis query and random values in the simulator)

This shouldn't be a problem for combinational cases or for >1 cycle
cases.

Why is it correct to pass in the inputs only when clk=0, though? This is
because the semantics of registers say that, when a register is at clk=1
and was previously at clk=0, they latch in the value from the previous
timestep when clk=0. So effectively, the value at the current timestep
does not matter. Thus, we never needed to be passing inputs at the clk=1
timestep!