Add Bitwuzla as a submodule (#365)
Closes #364 

Bitwuzla is under active development and there have been regressions
since we've been using it. To keep stuff from randomly breaking (as was
happening in #362) we should really tightly control the version of
Bitwuzla. Hence, we add it as a submodule. We also need to make sure we
use this submodule in the evaluation.