Encoding Lakeroad's IR in egglog (#354)
Experimenting with encoding Lakeroad's IR in egglog for eqsat + program
synthesis experiments originally proposed in my [Generals
talk](https://www.youtube.com/watch?v=M8i2HKEnoqI) (what I've been
calling Churchroad).

This PR includes a bunch of early experimentation. Experiments will
continue, but a lot of this is ready to be merged. It's likely that this
will be extracted to its own repository eventually; in that case, it'll
be good to have the link to the original PR:
https://github.com/uwsampl/lakeroad/pull/354

Thanks to @thiskappaisgrey for his contributions!

---------

Co-authored-by: Thanawat Techaumnuaiwit <thanatechaumnuaiwit@gmail.com>