Don't use oss cad suite (#425)
Closes #424 

Moving away from using oss-cad-suite. This has nothing to do with the
quality of oss-cad-suite -- in fact, I've had nothing but good
experiences using it. Here's the chain of reasoning that led me to
decide we should move away from using it.

1. I want to start releasing prebuilt Lakeroad.
2. Part of this involves packaging prebuilt solver binaries directly
into Rosette.
3. It's easier if those prebuilt solver binaries are statically linked,
because then we can just copy the executable by itself into Rosette.
4. That means we need statically-linked versions of the executables. 
5. oss-cad-suite's binaries are dynamically linked (makes sense,
oss-cad-suite is already very large)
6. Thus, we'd need to not use oss-cad-suite in our releasing script.
This would mean that the Docker image uses oss-cad-suite and the
releasing script doesn't.
7. I would like for the Docker and the release to be as similar as
possible.
8. Hence, I'm transitioning Docker away from oss-cad-suite.