Remove two tests
These tests no longer work with the Bitwuzla update.
I feel comfortable removing these tests.
Removing the first test is OK as it has a corresponding integration test.
That integration test will also have to be shifted away from using bitwuzla.
Removing the second test is OK as it is not actually a legal DSP case.